Move logging.py to dispatch_logging.py
