+++ docker-compose.yml
@@ -1,0 +1,31 @@
+version: "3.4"
+x-restart-policy: &restart_policy
+  restart: unless-stopped
+x-dispatch-defaults: &dispatch_defaults
+  <<: *restart_policy
+  env_file:
+    - .env
+  build:
+    # Pro-tip: point this to a relative directory containing the Dispatch
+    # project root to pick up changes from your dev environment
+    # (e.g., ../dispatch-root)
+    context: https://github.com/Netflix/dispatch.git#develop:.
+  image: dispatch-local
+  depends_on:
+    - postgres
+  env_file:
+    - .env
+services:
+  postgres:
+    <<: *restart_policy
+    env_file:
+      - .env
+    image: "postgres:9.6"
+    ports:
+      - "5432:5432"
+    volumes:
+      - "dispatch-postgres:/var/lib/postgresql/data"
+    env_file: ./.env
+volumes:
+  dispatch-postgres:
+    external: true
